update maya tool to compatible with maya 2022

+++ puzzle_maya/sbsar_mat_creation/controller/controller.py
@@ -7,7 +7,7 @@ import maya.cmds as cmds
 import sys
 
 mayaMainWindowPtr = omui.MQtUtil.mainWindow() 
-mayaMainWindow = wrapInstance(long(mayaMainWindowPtr), QtWidgets.QWidget)  # create a maya Main window
+mayaMainWindow = wrapInstance(int(mayaMainWindowPtr), QtWidgets.QWidget)  # create a maya Main window
 
 moduleImporterPath = 'general.modules_importer.modules_importer_function'
 importerFunction = None
